src/requirements.txt
beautifulsoup4==4.12.3
chardet==5.2.0
darkdetect==0.8.0
fpdf==1.7.2
Markdown==3.5.2
openpyxl==3.1.2
Pillow==10.2.0
psutil==5.9.6
pyinstaller==6.6.0
pystray==0.19.5
python_dateutil==2.8.2
python_docx==1.1.0
python_pptx==0.6.23
pywin32==306
Requests==2.32.3
ttkbootstrap==1.10.1
watchdog==3.0.0
winshell==0.6
